quick fix for alerts tab
